# openapi.yaml
openapi: 3.0.0
info:
  version: "1.0"
  title: "UniLinks Web API"
  description: "Web Resources Specification (A7) for UniLinks"
servers:
  - url: https://lbaw.fe.up.pt
    description: Production server
security: []
...
components:
  securitySchemes:
    cookieuser:
      type: apiKey
      in: cookie
      name: SESSIONID #cookie used to save session id
paths:
...
  /timeline:
    get:
      security:
        - cookieuser: []
      operationId: R302
      summary: "R302: View Feed"
      description: "Show the current logged user's feed. Access: Authenticated-USR, ADM"
      tags:
        - "M03"
      responses:
        "200":
          description: "Ok. Show feed"
...
  /api/group:
    delete:
      security:
        - cookieuser: []
      operationId: R508
      summary: "R508: Remove Member "
      description: "Remove member action. Access: Authenticated-USR, ADM"
      tags:
        - "M05"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                _token: #csrf
                  type: string
                userid:
                  type: integer
                groupid:
                  type: integer
              required:
                - _token
                - userid
                - groupid
      responses:
        "302":
          description: "Redirect after processing the friendship removal."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "The member removal was successfully done."
                  value: "/group/{group_id}"
                302Error:
                  description: "The member removal went wrong."
                  value: "/group/{group_id}"
...